requirement.txt
fastapi
uvicorn
httpx
biopython
python-dotenv
streamlit
requests